🎨 (config): Move config to top-level package and adjust imports
