+++ src/pyxon/storage/database/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from src.pyxon.config import Settings
+from src.config import Settings
 
 SQLALCHEMY_DATABASE_URL = Settings.DATABASE_URL
